clean up services - process captions services moved into job
