+++ app/jobs/process_caption_job.rb
@@ -1,22 +1,19 @@
-# post = InstaPost.first
-# result = ProcessCaptionService.new(post).call
-class ProcessCaptionService
+# insta_post = InstaPost.first
+# result = ProcessCaptionJob.perform_now(insta_post)
+class ProcessCaptionJob < ApplicationJob
+  queue_as :default
+
   delegate :link_to, to: 'ActionController::Base.helpers'
 
   include Rails.application.routes.url_helpers
 
-  attr_reader :post
+  def perform(insta_post)
+    @insta_post = insta_post
+    return if @insta_post.caption.blank?
 
-  def initialize(post)
-    @post = post
-  end
-
-  def call
-    return if post.caption.blank?
-
-    processed_caption = with_regex(post.caption)
-    post.reload
-    post.update!(processed_caption:)
+    processed_caption = with_regex(@insta_post.caption)
+    @insta_post.reload
+    @insta_post.update!(processed_caption:)
   end
 
   private
@@ -36,7 +33,7 @@ class ProcessCaptionService
     results.flatten.each do |result|
       result_link =
         link_to result,
-                insta_user_posts_path(post.insta_user, caption: result, onlypath: true),
+                insta_user_posts_path(@insta_post.insta_user, caption: result, onlypath: true),
                 data: { turbo: false },
                 class: css_class
       text.gsub!(result, result_link)
