--- app/jobs/process_caption_service_job.rb
@@ -1,7 +1,0 @@
-class ProcessCaptionServiceJob < ApplicationJob
-  queue_as :default
-
-  def perform(insta_post)
-    ProcessCaptionService.new(insta_post).call
-  end
-end
